refactor(response): unify stream handling and polish multi-model UX
  - centralize chat and single-model responders under core/response using the new
    stream runner + responder factory to eliminate duplicated logic
  - add core/response/stream-runner.js and handlers.js to share AbortController,
    message prep, and spinner orchestration across responders
  - update multi-model executor to reuse the shared runner, keeping winner logic
    while removing bespoke stream plumbing
  - fix winner/queued-model race by caching the waiting message until the leader
    finishes and then restarting the spinner cleanly
  - update bin/app.js imports and refresh AGENTS.md to document the new layout
