+++ bin/app.js
@@ -4,10 +4,10 @@ import { outputHandler } from '../core/print/index.js'
 import { errorHandler } from '../core/error-system/index.js'
 import { createRouter } from '../core/router.js'
 import { createApplicationLoop } from '../core/application-loop/index.js'
-import { createChatHandler } from '../core/chat-handler.js'
 import {
   createSingleModelCommand,
   createMultiModelCommand,
+  createChatHandler,
 } from '../core/response/index.js'
 import { systemCommandHandler } from '../core/system-command-handler.js'
 import { getStateManager, stateManagerEvents } from '../core/StateManager.js'
