+++ core/response/index.js
@@ -1,3 +1,5 @@
 export { createResponseSessionFactory } from './session.js'
-export { createSingleModelCommand } from './single-model-command.js'
+export { createStreamCommandRunner } from './stream-runner.js'
+export { createSingleModelCommand } from './single.js'
 export { createMultiModelCommand } from './multi-model/index.js'
+export { createChatHandler } from './chat.js'
